Tidy sumario from xml file

Cleans in a tidy format the sumario file

Usage

tidy_sumario(x)

Arguments

x

A XML file.

Value

A data.frame with one line for each publication. Including date, sumario number, section, section text, departament, departament number, epigraph, brief text, publication code and number of pages of the pdf.

Examples

id <- sumario_nbo(format(as.Date("2017/10/02", "%Y/%m/%d"), "%Y%m%d"))
sumario_file <- get_xml(query_xml(id))
m <- tidy_sumario(sumario_file)

head(m)
#>         date sumario_nbo   sumario_code section
#> 1 2017-10-02         237 BOE-S-2017-237       1
#> 2 2017-10-02         237 BOE-S-2017-237       1
#> 3 2017-10-02         237 BOE-S-2017-237      2A
#> 4 2017-10-02         237 BOE-S-2017-237      2A
#> 5 2017-10-02         237 BOE-S-2017-237      2A
#> 6 2017-10-02         237 BOE-S-2017-237      2A
#>                                                              section_number
#> 1                                                I. Disposiciones generales
#> 2                                                I. Disposiciones generales
#> 3 II. Autoridades y personal. - A. Nombramientos, situaciones e incidencias
#> 4 II. Autoridades y personal. - A. Nombramientos, situaciones e incidencias
#> 5 II. Autoridades y personal. - A. Nombramientos, situaciones e incidencias
#> 6 II. Autoridades y personal. - A. Nombramientos, situaciones e incidencias
#>                                         departament departament_etq
#> 1                               JEFATURA DEL ESTADO            7723
#> 2 MINISTERIO DE ASUNTOS EXTERIORES Y DE COOPERACIÓN            7120
#> 3                CONSEJO GENERAL DEL PODER JUDICIAL            1820
#> 4                CONSEJO GENERAL DEL PODER JUDICIAL            1820
#> 5                CONSEJO GENERAL DEL PODER JUDICIAL            1820
#> 6                CONSEJO GENERAL DEL PODER JUDICIAL            1820
#>                   epigraph
#> 1 Acuerdos internacionales
#> 2 Acuerdos internacionales
#> 3            Nombramientos
#> 4                 Destinos
#> 5              Situaciones
#> 6              Situaciones
#>                                                                                                                                                                                                                                          text
#> 1                                                                                     Instrumento de ratificación del Protocolo Adicional al Convenio sobre traslado de personas condenadas, hecho en Estrasburgo el 18 de diciembre de 1997.
#> 2                                                Enmiendas a los Anexos 2, 6 y 7 del Convenio Aduanero relativo al Transporte Internacional de Mercancías al amparo de los Cuadernos TIR, adoptadas en Ginebra el 10 y 11 de febrero de 2016.
#> 3                                                                                    Real Decreto 847/2017, de 15 de septiembre, por el que se nombra a la magistrada doña María Gloria Pérez Padilla, Juez de lo Penal número 3 de Sabadell.
#> 4                 Real Decreto 846/2017, de 15 de septiembre, por el que se destina a los Magistrados que se relaciona, como consecuencia del concurso resuelto por acuerdo de la Comisión Permanente del Consejo General del Poder Judicial.
#> 5                                         Acuerdo de 6 de julio de 2017, de la Comisión Permanente del Consejo General del Poder Judicial, por el que se declara la jubilación voluntaria anticipada de doña Juliana Mercedes Gómez Ferreras.
#> 6 Acuerdo de 21 de septiembre de 2017, de la Comisión Permanente del Consejo General del Poder Judicial, por el que queda a disposición del Presidente del Tribunal Superior de Justicia de Madrid, la Magistrada doña Carmen Huerta Manzano.
#>        publication pages
#> 1 BOE-A-2017-11117    10
#> 2 BOE-A-2017-11118    16
#> 3 BOE-A-2017-11120     1
#> 4 BOE-A-2017-11119     3
#> 5 BOE-A-2017-11121     1
#> 6 BOE-A-2017-11122     1
tail(m)
#>           date sumario_nbo   sumario_code section
#> 408 2017-10-02         237 BOE-S-2017-237      5B
#> 409 2017-10-02         237 BOE-S-2017-237      5C
#> 410 2017-10-02         237 BOE-S-2017-237      5C
#> 411 2017-10-02         237 BOE-S-2017-237      5C
#> 412 2017-10-02         237 BOE-S-2017-237      5C
#> 413 2017-10-02         237 BOE-S-2017-237      5C
#>                                 section_number           departament
#> 408 V. Anuncios. - B. Otros anuncios oficiales         UNIVERSIDADES
#> 409    V. Anuncios. - C. Anuncios particulares ANUNCIOS PARTICULARES
#> 410    V. Anuncios. - C. Anuncios particulares ANUNCIOS PARTICULARES
#> 411    V. Anuncios. - C. Anuncios particulares ANUNCIOS PARTICULARES
#> 412    V. Anuncios. - C. Anuncios particulares ANUNCIOS PARTICULARES
#> 413    V. Anuncios. - C. Anuncios particulares ANUNCIOS PARTICULARES
#>     departament_etq epigraph
#> 408             310     <NA>
#> 409            9202     <NA>
#> 410            9202     <NA>
#> 411            9202     <NA>
#> 412            9202     <NA>
#> 413            9202     <NA>
#>                                                                           text
#> 408 Anuncio de la Universidad de Cádiz sobre extravío de título universitario.
#> 409                                             FUNDACIÓN AMANCIO ORTEGA GAONA
#> 410                                       ILUSTRE COLEGIO NOTARIAL DE VALENCIA
#> 411                                            NOTARÍA DE CARLOS ALCOCER TORRA
#> 412                                             NOTARÍA DE JOSÉ LUIS RUIZ ABAD
#> 413                                      NOTARÍA DE MERCEDES ALCÁZAR RODRÍGUEZ
#>          publication pages
#> 408 BOE-B-2017-56825     1
#> 409 BOE-B-2017-56826     1
#> 410 BOE-B-2017-56827     1
#> 411 BOE-B-2017-56828     1
#> 412 BOE-B-2017-56829     1
#> 413 BOE-B-2017-56830     1

Site built with pkgdown 2.0.6.

Template rostemplate by dieghernan, based on Bootstrapious.